Address review: EnvLocalFile class with rollback, 1:1 handler files, drop redundant gitignore write, standardize testIO stdin
