+++ src/testing/testIO.tsx
@@ -16,6 +16,9 @@ export interface TestIO {
 
 export interface TestIOOptions {
   isTTY?: boolean;
+  // stdin pre-fills the input stream with piped content, then ends it. Use this
+  // to test handlers that read a secret or prompt from stdin ('-').
+  stdin?: string;
 }
 
 // collect wraps a PassThrough, accumulating everything written to it as a string.
@@ -33,10 +36,12 @@ function collect(): { stream: NodeJS.WriteStream; read: () => string } {
 
 // testIO builds a fresh in-memory TestIO for a single test. stdin is an idle
 // PassThrough (no input) so screens that read input simply see nothing.
-export function testIO({ isTTY = false }: TestIOOptions = {}): TestIO {
+export function testIO({ isTTY = false, stdin: stdinContent }: TestIOOptions = {}): TestIO {
   const out = collect();
   const err = collect();
-  const stdin = new PassThrough() as unknown as NodeJS.ReadStream;
+  const stdinStream = new PassThrough();
+  if (stdinContent !== undefined) stdinStream.end(stdinContent);
+  const stdin = stdinStream as unknown as NodeJS.ReadStream;
 
   for (const stream of [stdin, out.stream, err.stream]) {
     Object.defineProperty(stream, "isTTY", { configurable: true, value: isTTY });
